+++ R/LearnerDensityHist.R
@@ -1,1 +1,0 @@
-# LearnerDensityHist <- R6::R6Class("LearnerDensityHist", inherit = LearnerDensity)