+++ R/zzz.R
@@ -1,9 +1,10 @@
-#' @import data.table
 #' @import checkmate
-#' @import paradox
+#' @import data.table
+#' @import distr6
 #' @import mlr3
 #' @import mlr3misc
 #' @import mlr3pipelines
+#' @import paradox
 #' @importFrom R6 R6Class
 #' @importFrom utils data head tail
 #' @importFrom stats reformulate model.matrix model.frame sd
@@ -19,21 +20,11 @@ register_mlr3 = function() {
     x$task_types = setkeyv(rbind(x$task_types, rowwise_table(
       ~type,     ~package,    ~task,         ~learner,         ~prediction,         ~measure,
       "surv",    "mlr3proba", "TaskSurv",    "LearnerSurv",    "PredictionSurv",    "MeasureSurv"
-      # "density", "mlr3proba", "TaskDensity", "LearnerDensity", "PredictionDensity", "MeasureDensity",
-      # "probreg", "mlr3proba", "TaskProbreg", "LearnerProbreg", "PredictionProbreg", "MeasureDensity"
     )), "type")
     x$task_col_roles$surv = c("feature", "target", "label", "order", "group", "weight", "stratum")
-    # x$task_col_roles$density = c("feature", "target", "label", "order", "group", "weight")
-    # x$task_col_roles$probreg = c("feature", "target", "label", "order", "group", "weight")
     x$task_properties$surv = c("weights", "groups")
-    # x$task_properties$density = c("weights", "groups")
-    # x$task_properties$probreg = c("weights", "groups")
     x$learner_properties$surv = x$learner_properties$regr
-    # x$learner_properties$density = x$learner_properties$regr
-    # x$learner_properties$probreg = x$learner_properties$regr
     x$measure_properties$surv = x$measure_properties$regr
-    # x$measure_properties$density = x$measure_properties$regr
-    # x$measure_properties$probreg = x$measure_properties$regr
     x$learner_predict_types$surv = list(crank = c("crank","lp","distr","response"),
                                         distr = c("crank","lp","distr","response"),
                                         lp = c("crank","lp","distr","response"),
@@ -41,22 +32,37 @@ register_mlr3 = function() {
     x$default_measures$surv = "surv.harrellC"
   }
 
+  if (!grepl("dens", x$task_types[,"type"])) {
+     x = utils::getFromNamespace("mlr_reflections", ns = "mlr3")
+     x$task_types = setkeyv(rbind(x$task_types, rowwise_table(
+        ~type,     ~package,    ~task,         ~learner,         ~prediction,         ~measure,
+        "dens",    "mlr3proba", "TaskDens",    "LearnerDens",    "PredictionDens",    "MeasureDens"
+     )), "type")
+     x$task_col_roles$dens = c("feature", "target", "label", "order", "group", "weight", "stratum")
+     x$task_properties$dens = c("weights", "groups")
+     x$learner_properties$dens = x$learner_properties$regr
+     x$measure_properties$dens = x$measure_properties$regr
+     x$learner_predict_types$dens = list(pdf = c("pdf","cdf"),
+                                         cdf = c("pdf","cdf"))
+     x$default_measures$dens = "dens.logloss"
+  }
+
   # tasks
    x = utils::getFromNamespace("mlr_tasks", ns = "mlr3")
-   # x$add("precip", load_task_precip)
+   x$add("precip", load_precip)
    x$add("rats", load_rats)
    x$add("lung", load_lung)
    x$add("unemployment", load_unemployment)
 
   # generators
    x = utils::getFromNamespace("mlr_task_generators", ns = "mlr3")
-   # x$add("friedman1dens", TaskGeneratorFriedman1Dens)
+   x$add("simdens", TaskGeneratorSimdens)
    x$add("simsurv", TaskGeneratorSimsurv)
 
   # learners
    x = utils::getFromNamespace("mlr_learners", ns = "mlr3")
-   # x$add("density.kde", LearnerDensityKDE)
-   # x$add("probreg.gaussian", LearnerProbregGaussian)
+
+   x$add("dens.hist", LearnerDensHistogram)
 
    x$add("surv.coxph", LearnerSurvCoxPH)
    x$add("surv.kaplan", LearnerSurvKaplan)
@@ -79,8 +85,11 @@ register_mlr3 = function() {
 
   # measures
    x = utils::getFromNamespace("mlr_measures", ns = "mlr3")
-   # x$add("density.logloss", MeasureDensityLogloss)
+
+   x$add("dens.logloss", MeasureDensLogloss)
+
    # x$add("regr.logloss", MeasureRegrLogloss)
+
    x$add("surv.graf", MeasureSurvGraf)
    x$add("surv.grafSE", MeasureSurvGrafSE)
    x$add("surv.logloss", MeasureSurvLogloss)