+++ tests/testthat/test_mlr_learners_density_kde.R
@@ -1,8 +1,0 @@
-# context("density.kde")
-#
-# test_that("autotest", {
-#   learner = lrn("density.kde")
-#   expect_learner(learner)
-#   result = run_autotest(learner)
-#   expect_true(result, info = result$error)
-# })