+++ NEWS.md
@@ -1,3 +1,8 @@
+# mlr3proba 0.1.4
+
+* Density estimation has now been added to mlr3proba, see `TaskDens`, `LearnerDens`, `PredictionDens`, and `MeasureDens`.
+* Added learners for density estimation, see `mlr3::mlr_learners$keys("^dens")` for the full list
+
 # mlr3proba 0.1.3
 
 * Changed `lp` in `surv.parametric` to include the intercept, which is in line with `survival::survreg`. Now `exp(pred$lp)` is equal to the predicted survival time for AFTs